src/data/data_generator.py
import numpy as np
import pandas as pd
import yaml
from pathlib import Path
from typing import Dict, Tuple, List
import json

class SteelCastingDataGenerator:
    """Generates synthetic steel casting process data"""
    
    def __init__(self, config_path: str):
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)
        self.random_state = np.random.RandomState(
            self.config['data_generation']['random_seed']
        )
    
    def generate_cast_sequence(self, cast_id: str) -> Tuple[pd.DataFrame, Dict]:
        """
        Generate time series data for a single cast.
        
        Parameters:
            cast_id (str): Unique identifier for the cast.
        
        Returns:
            Tuple[pd.DataFrame, Dict]: A tuple containing:
                - pd.DataFrame: Time series data for the cast.
                - Dict: Metadata associated with the cast.
        
        Implementation Overview:
            This method generates synthetic time series data for a steel casting process
            based on the configuration provided during initialization. The data includes
            process parameters (e.g., temperature, pressure) and timestamps.
        """
        # TODO: Implement cast sequence generation
        pass
    
    def generate_dataset(self) -> None:
        """
        Generate a complete synthetic dataset for steel casting processes.
        
        This method is intended to create a dataset containing synthetic time-series data
        for multiple steel casting sequences. The dataset will be generated based on the
        configuration parameters provided during the initialization of the class.
        
        Workflow:
        - Iterate over a predefined number of casting sequences.
        - For each sequence, call the `generate_cast_sequence` method to generate time-series data.
        - Aggregate the data into a single dataset.
        - Save the dataset to a file or return it as a DataFrame.
        
        Expected Outputs:
        - A complete synthetic dataset in the form of a Pandas DataFrame or saved to a file.
        
        Note:
        This method is currently a placeholder and requires implementation.
        """
        # TODO: Implement dataset generation
        pass