src/features/feature_engineer.py
import pandas as pd
import numpy as np
from typing import Dict, List
from sklearn.preprocessing import StandardScaler

class CastingFeatureEngineer:
    """Extract features for baseline model"""
    
    def __init__(self):
        self.scaler = StandardScaler()
        self.feature_columns = []
    
    def extract_statistical_features(self, df: pd.DataFrame) -> Dict:
        """
        Extract mean, std, min, max, median for each sensor.
        
        Args:
            df (pd.DataFrame): Time series data for a single cast
            
        Returns:
            Dict: Statistical features for each sensor
        """
        # TODO: Implement statistical feature extraction
        pass
    
    def extract_stability_features(self, df: pd.DataFrame) -> Dict:
        """
        Extract spike counts, excursion frequencies.
        
        Args:
            df (pd.DataFrame): Time series data for a single cast
            
        Returns:
            Dict: Stability features
        """
        # TODO: Implement stability feature extraction
        pass
    
    def extract_duration_features(self, df: pd.DataFrame) -> Dict:
        """
        Extract time spent at extremes.
        
        Args:
            df (pd.DataFrame): Time series data for a single cast
            
        Returns:
            Dict: Duration-based features
        """
        # TODO: Implement duration feature extraction
        pass
    
    def extract_interaction_features(self, features: Dict) -> Dict:
        """
        Create cross-sensor interaction features.
        
        Args:
            features (Dict): Existing features
            
        Returns:
            Dict: Interaction features
        """
        # TODO: Implement interaction feature creation
        pass
    
    def transform_cast(self, time_series: pd.DataFrame) -> Dict:
        """
        Transform single cast to feature vector.
        
        Args:
            time_series (pd.DataFrame): Time series data for a single cast
            
        Returns:
            Dict: Complete feature vector for the cast
        """
        # TODO: Implement complete feature transformation
        pass
    
    def fit_scaler(self, feature_matrix: pd.DataFrame) -> None:
        """
        Fit feature scaler on training data.
        
        Args:
            feature_matrix (pd.DataFrame): Training features
        """
        # TODO: Implement scaler fitting
        pass
    
    def scale_features(self, features: pd.DataFrame) -> pd.DataFrame:
        """
        Scale features using fitted scaler.
        
        Args:
            features (pd.DataFrame): Features to scale
            
        Returns:
            pd.DataFrame: Scaled features
        """
        # TODO: Implement feature scaling
        pass